% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment80.R
\name{comment80}
\alias{comment80}
\title{comment80}
\usage{
comment80(textString)
}
\arguments{
\item{textString}{A comment text string}
}
\value{
A comment printed to the console
}
\description{
comment80 prints to the console a comment of 80 width with the text string centered and split evently between lines
}
\examples{
#comment80(textString="This is a short comment")
#comment80(textString="This is a longer text string that will end up having to be split between two different lines. It will be split such that the lines are approximately equal.")
}

