% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentProjectInfo.R
\name{commentProjectInfo}
\alias{commentProjectInfo}
\title{commentProjectInfo}
\usage{
commentProjectInfo(ProjectName = "", Researcher = "", Statistician = "",
  StartDate = "", UpdateDate = "")
}
\arguments{
\item{ProjectName}{Name of the current project}

\item{Researcher}{Name of the researcher/client}

\item{Statistician}{Name of the statistician/programmer}

\item{StartDate}{Start date}

\item{UpdateDate}{Updated date}
}
\value{
A comment printed to the console
}
\description{
commentProjectInfo prints a comment to the console that includes relevant project information.
}
\examples{
#commentProjectInfo(ProjectName="This is the project name",
#                  Researcher = 'John Doe',
#                  Statistician = 'Andrew Taylor',
#                  StartDate = 'December 9, 2015',
#                 UpdateDate = '')
}

