% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProject.R
\name{newProject}
\alias{newProject}
\title{newProject}
\usage{
newProject(Dir, FolderName, addReportSkeleton = F, recursive = FALSE)
}
\arguments{
\item{Dir}{Path pointing to where the new project directory should be added}

\item{FolderName}{Name of the new directory folder}

\item{addReportSkeleton}{If TRUE, an RMarkdown skeleton is added to /Reports. Note: This functionality will only work for Andrew Taylor, as the report skeleton is in my directory. If people want to use this, the report skeletons can be moved to the common M:/bioteam/ directory.}

\item{recursive}{If TRUE, if the final folder in `Dir` does not yet exist, it will be created}
}
\description{
newProject creates a new project directory folder with 'Data', 'Documents', 'Output', 'Reports', and 'Syntax' subfolders.
}
\examples{
#newProject("C:/User/Me/", "NewProject")
}

