% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwdDate.R
\name{setwdDate}
\alias{setwdDate}
\title{setwdDate}
\usage{
setwdDate(path)
}
\arguments{
\item{path}{A path to the directory in which the newly created folder will be appended to.}
}
\description{
setwdDate creates a folder with the current day's date (YYYYMMDD) in a given directory, and then sets the working directory to that folder.
}
\examples{
#setwdDate("C:/MyFakePath/Output/")
##Would create a folder:
#paste0("C:/MyFakePath/Output/", format(Sys.Date(), "\%Y\%m\%d"))
##And would set the working directory to that folder
}

